requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
Deprecated==1.2.13
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
dnspython==2.1.0
idna==3.2
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.2.0
pytz==2021.3
requests==2.26.0
Rx==3.2.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
vonage==2.5.5
wrapt==1.13.1
